+++ qupath-core/src/main/java/qupath/lib/io/PathObjectIO.java
@@ -1,4 +1,27 @@
-package qupath.lib.objects;
+/*-
+ * #%L
+ * This file is part of QuPath.
+ * %%
+ * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
+ * Contact: IP Management (ipmanagement@qub.ac.uk)
+ * Copyright (C) 2018 - 2021 QuPath developers, The University of Edinburgh
+ * %%
+ * QuPath is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * QuPath is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License 
+ * along with QuPath.  If not, see <https://www.gnu.org/licenses/>.
+ * #L%
+ */
+
+package qupath.lib.io;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -26,7 +49,8 @@ import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 
 import qupath.lib.common.GeneralTools;
-import qupath.lib.io.GsonTools;
+import qupath.lib.objects.PathObject;
+import qupath.lib.objects.PathObjectTools;
 
 /**
  * Class to export {@code PathObject}s in GeoJSON format.
@@ -180,13 +204,4 @@ public final class PathObjectIO {
 		return exts;
 	}
 
-	/**
-	 * Return whether the {@code PathObject} is an ellipse.
-	 * 
-	 * @param ann
-	 * @return isEllipse
-	 */
-	public static boolean isEllipse(PathObject ann) {
-		return ann.getROI() != null && ann.getROI().getRoiName().equals("Ellipse");
-	}
 }