+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/ExportObjectsCommand.java
@@ -1,3 +1,26 @@
+/*-
+ * #%L
+ * This file is part of QuPath.
+ * %%
+ * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
+ * Contact: IP Management (ipmanagement@qub.ac.uk)
+ * Copyright (C) 2018 - 2021 QuPath developers, The University of Edinburgh
+ * %%
+ * QuPath is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * QuPath is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License 
+ * along with QuPath.  If not, see <https://www.gnu.org/licenses/>.
+ * #L%
+ */
+
 package qupath.lib.gui.commands;
 
 import java.io.File;
@@ -18,9 +41,9 @@ import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.dialogs.Dialogs;
 import qupath.lib.gui.tools.PaneTools;
+import qupath.lib.io.PathObjectIO;
 import qupath.lib.objects.PathObject;
 import qupath.lib.objects.hierarchy.PathObjectHierarchy;
-import qupath.lib.objects.PathObjectIO;
 import qupath.lib.plugins.workflow.DefaultScriptableWorkflowStep;
 
 /**
@@ -90,7 +113,7 @@ public final class ExportObjectsCommand {
 		toProcess = toProcess.stream().filter(e -> !e.isRootObject()).collect(Collectors.toList());
 
 		// Check if includes ellipse(s), as they will need to be polygonized
-		var nEllipses = toProcess.stream().filter(ann -> PathObjectIO.isEllipse(ann)).count();
+		var nEllipses = toProcess.stream().filter(ann -> isEllipse(ann)).count();
 		if (nEllipses > 0) {
 			var response = Dialogs.showYesNoDialog("Ellipse polygonization", String.format("%d ellipse(s) will be polygonized, continue?", nEllipses));
 			if (!response)
@@ -138,4 +161,15 @@ public final class ExportObjectsCommand {
 		historyWorkflow.addStep(new DefaultScriptableWorkflowStep(methodTitle, map, methodString));		
 		return true;
 	}
+	
+	/**
+	 * Return whether the {@code PathObject} is an ellipse.
+	 * 
+	 * @param ann
+	 * @return isEllipse
+	 */
+	private static boolean isEllipse(PathObject ann) {
+		return ann.getROI() != null && ann.getROI().getRoiName().equals("Ellipse");
+	}
+	
 }